requirements.txt
Django>=4.0,<5.0
djangorestframework
requests
psycopg2-binary
python-decouple
django-cors-headers